refactor: Move csgo based crates into their own package
