+++ paper/src/main/java/com/badbones69/crazycrates/api/builders/CrateBuilder.java
@@ -22,7 +22,6 @@ import com.badbones69.crazycrates.CrazyCrates;
 import com.badbones69.crazycrates.api.events.CrateOpenEvent;
 import com.badbones69.crazycrates.utils.MiscUtils;
 import java.util.List;
-import java.util.TimerTask;
 
 public abstract class CrateBuilder extends BukkitRunnable {
