+++ paper/src/main/java/com/badbones69/crazycrates/managers/crates/CrateManager.java
@@ -5,7 +5,8 @@ import com.badbones69.crazycrates.api.FileManager;
 import com.badbones69.crazycrates.api.FileManager.Files;
 import com.badbones69.crazycrates.api.enums.BrokeLocation;
 import com.badbones69.crazycrates.managers.crates.types.*;
-import com.badbones69.crazycrates.managers.crates.types.animation.RouletteTriple;
+import com.badbones69.crazycrates.managers.crates.types.csgo.TripleAnimation;
+import com.badbones69.crazycrates.managers.crates.types.csgo.SingleAnimation;
 import org.bukkit.inventory.meta.ItemMeta;
 import org.bukkit.permissions.Permission;
 import org.bukkit.permissions.PermissionDefault;
@@ -363,8 +364,8 @@ public class CrateManager {
         CrateBuilder crateBuilder;
 
         switch (crate.getCrateType()) {
-            case csgo -> crateBuilder = new CsgoCrate(crate, player, 27);
-            case csgo_triple -> crateBuilder = new RouletteTriple(crate, player, 54);
+            case csgo -> crateBuilder = new SingleAnimation(crate, player, 27);
+            case csgo_triple -> crateBuilder = new TripleAnimation(crate, player, 54);
             case wonder -> crateBuilder = new WonderCrate(crate, player, 45);
             case wheel -> crateBuilder = new WheelCrate(crate, player, 54);
             case roulette -> crateBuilder = new RouletteCrate(crate, player, 45);
@@ -419,7 +420,7 @@ public class CrateManager {
                 crateBuilder = new QuickCrate(crate, player, location);
             }
             default -> {
-                crateBuilder = new CsgoCrate(crate, player, 27);
+                crateBuilder = new SingleAnimation(crate, player, 27);
 
                 if (this.plugin.isLogging()) {
                     List.of(
